Initial work on supplier entries page
