+++ locales/en/gate.json
@@ -1,5 +1,6 @@
 {
     "gate" : "Gate",
+    "entries": "Entry management",
     "desc": "Manage gate entrance",
     "search": "Search by ticket number / person name / email / id",
     "enter_ticket": "Enter ticket",
