+++ locales/he/gate.json
@@ -1,5 +1,6 @@
 {
     "gate" : "שער",
+    "entries": "ניהול כניסות",
     "desc": "ניהול כניסות לאירוע",
     "search": "חיפוש באמצעות שם פרטי / שם משפחה / שם מלא / דואר אלקטרוני / מספר כרטיס / תעודת זהות",
     "enter_ticket": "כניסה לאירוע",
