+++ routes/index.js
@@ -13,7 +13,6 @@ app.use("/:lng/npo", require("./pages/npo_routes"));
 app.use("/:lng/npo-admin", require("./pages/npo_admin_routes"));
 app.use("/:lng/gate", require("./pages/gate_routes"));
 app.use("/:lng/volunteering", require("./pages/volunteering_routes"));
-app.use('/:lng/supplier-entries', require('./pages/supplier_entries_routes'));
 app.use('/:lng/', require('./pages/suppliers_routes'));
 
 //TODO: refactor camps routes
