+++ views/pages/suppliers/supplier-entries.jade
@@ -1,14 +1,6 @@
 extends ../../includes/page
 block content
-    .camps.camp_admin_index(ng-app="supplierEntriesController")
-        script.
-            var groups_prototype="#{__groups_prototype}";            
-        if controllDates
-            script.
-                const controllDates = {
-                    appreciation_tickets_allocation_start : '#{controllDates.appreciation_tickets_allocation_start}',
-                    appreciation_tickets_allocation_end : '#{controllDates.appreciation_tickets_allocation_end}',
-                }ת
+    .camps.camp_admin_index(ng-app="ngSuppliers")
         .container.col-md-12
             h1=t(t_prefix+'entries.suppliers')
             p.lead.text-muted=t(t_prefix+'entries.desc')
@@ -17,13 +9,13 @@ block content
                 div.block-center
                     div.input-group
                         if isAdmin
-                            a.Btn.Btn__green.card-switcher#4(href=`/${language}/supplier-entries/new-entrance`, role='button')
+                            a.Btn.Btn__green.card-switcher#4(role='button')
                                 span.glyphicon.glyphicon-plus(aria-hidden="true")
                                 span=t(t_prefix+'entries.supplier_entrance')
-            
+
         .cards--wrapper.card__shad
             //- Card 1
-            .card.card-first.card__padding--even.col-xs-12.table-responsive(ng-controller="manageSupplierEntriesController")
+            .card.card-first.card__padding--even.col-xs-12.table-responsive(ng-controller="supplierEntriesController")
                 h4=t(t_prefix+'edit_new.suppliers')
                 .controls.row
                   .col-xs-12.col-md-12
@@ -43,39 +35,39 @@ block content
                     tbody
                         tr(ng-repeat="supplier in suppliers | filter: searchCamp | orderBy: orderCamps")
                             td {{ supplier.supplier_id }}
-                            td 
+                            td
                                 //- a(ng-href=`/${language}/suppliers/{{supplier.supplier_id}}`)
                                 //-     span {{supplier.supplier_name_en}}
-                                //- br 
+                                //- br
                                 //- a(ng-href=`/${language}/suppliers/{{supplier.supplier_id}}`)
                                 //- TODO link name to supplier page
-                                a(ng-href="#") 
+                                a(ng-href="#")
                                     span {{supplier.supplier_name_en}}
-                                br 
+                                br
                                 a(ng-href="#")
                                     span {{supplier.supplier_name_he}}
                             td {{ supplier.main_contact_name }}
                             td {{ supplier.main_contact_position }}
                             td {{ supplier.main_contact_phone_number }}
                             td {{ supplier.updated_at }}
-                            td {{ supplier.supplier_website_link }}                            
+                            td {{ supplier.supplier_website_link }}
                             td {{ supplier.supplier_midmarket_link }}
                             td {{ supplier.comments }}
                             td {{ supplier.supplier_category }}
                             td
                                 if (user.isAdmin || (user.isCampsAdmin && (!controllDates.edit_camps_lastDate || new Date() < controllDates.edit_camps_lastDate)))
                                     a.cursor-p(ng-href=`/${language}/suppliers/{{supplier.supplier_id}}` + '/edit') &nbsp;
-                                        button.btn.btn-xs 
+                                        button.btn.btn-xs
                                             span(class='glyphicon glyphicon-pencil')
                                             =t(t_prefix+'edit_new.edit')
             .choose_name.card.card-forth.card__padding--even.card-hide.col-xs-12
-                h2=t(t_prefix+'edit_new.new_supplier')
+                h2=t(t_prefix+'edit_new.search')
                 h4=t(t_prefix+'edit_new.id_choosing_title')
                 .col-sm-8.col-xs-12
                     .pos-r
                         label=t(t_prefix+'edit_new.id')
-                        input(id='new-supplier-button' type="number", class="form-control", name="supplier_id", autofocus="true", required)
-                        span.indicator
-                            span.glyphicon
+                        input(id='new-supplier-button' type="text", class="form-control", name="supplier_id", autofocus="true", required)
                 .col-sm-4.col-xs-12
                     a.Btn.Btn__green.Btn__inline.btn.disabled(id="check_supplier_id", role='button')=t(t_prefix+'edit_new.btn_continue')
+                    a.Btn.Btn__red.Btn__inline.btn.card-switcher#1(role='button')
+                        span=t(t_prefix+'entries.cancel')
